requirements.txt
# FastAPI et serveur web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Base de données
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Sécurité et authentification
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Configuration et environnement
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# CORS et middleware
python-dateutil==2.8.2

# Supabase (optionnel si besoin du client)
supabase==2.3.1

# Utilitaires
email-validator==2.1.0

# Production
gunicorn==21.2.0

# Monitoring et logs (optionnel)
# sentry-sdk==1.39.2
